requirements.txt
tgcrypto
pyrofork
aiohttp
psutil
python-ffmpeg
pornhub-api==0.2.0
youtube-dl==2021.6.6
Pillow
pytz
dnspython
hachoir
humanize
ffmpeg-python
motor
beautifulsoup4
html5lib
aiofiles
psutil
moviepy
cloudscraper
img2pdf
python-dateutil
spidermonkey